pyproject.toml
[project]
name = "pr_agent"
version = "0.1.0"
description = "Autonomous code review agent system that uses AI to analyze GitHub pull requests"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.4.0",
    "fastapi[standard]>=0.115.5",
    "langchain-openai>=0.2.10",
    "langchain>=0.3.8",
    "openai>=1.55.1",
    "pygithub>=2.5.0",
    "python-dotenv>=1.0.1",
    "redis>=5.2.0",
]
